# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Duka POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, paid, cancelled, failed]
      responses:
        "200":
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order from cart items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order and restore stock if it was paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order permanently (restores stock if it was paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: An order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        priceCents:
          type: integer
          description: Price in KES cents
        stock:
          type: integer
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sku
        - description
        - priceCents
        - stock
        - category
        - imageUrl
        - createdAt
        - updatedAt
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        priceCents:
          type: integer
          minimum: 0
        stock:
          type: integer
          minimum: 0
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - name
        - priceCents
        - stock
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        priceCents:
          type: integer
          minimum: 0
        stock:
          type: integer
          minimum: 0
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        lineTotalCents:
          type: integer
      required:
        - id
        - orderId
        - productId
        - productName
        - quantity
        - unitPriceCents
        - lineTotalCents
    Order:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        status:
          type: string
          enum: [pending, paid, cancelled, failed]
        totalCents:
          type: integer
        customerPhone:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        cashierName:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [mpesa, cash]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - reference
        - status
        - totalCents
        - customerPhone
        - customerName
        - cashierName
        - paymentMethod
        - items
        - createdAt
        - updatedAt
    CreateOrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
      required:
        - productId
        - quantity
    CreateOrderInput:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOrderItemInput"
        customerPhone:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        cashierName:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [mpesa, cash]
      required:
        - items
        - paymentMethod
...