# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duka POS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  reference: string;
  status: OrderStatus;
  totalCents: number;
  /** @nullable */
  customerPhone: string | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  cashierName: string | null;
  paymentMethod: OrderPaymentMethod;
  items: OrderItem[];
  createdAt: Date;
  updatedAt: Date;
}